==> src/validation/02.1_insilico_digest_prepare_genomes_and_tracks.ipynb <==
## Insilico digest preparation
Download of the reference data from the UCSC
Author: Daria Romanovskaia dromanovskaia@cemm.at  
Date: 22.12.2020 (upd. 1.02.2021, upd 18.10.2021)

source(file.path(Sys.getenv("CODEBASE"),"DNAmeth500species/src/00.0_init.R"))

library(rtracklayer)

## Set of species to analyse:  

We want to run the insilico RRBS on all the genomes, we used for the cross-mapping and expand by other possible maches.We have defined a set of speices, that are of potential interest for us as they have avaliable genomes + needed annotation tracks.
First version of the validation_species contained only the species name and the class and was further annotated along this notebook

getwd()

species from cross-mapping:

df_crossmap <- fread(file.path(analysis_dir,"validation", "01_crossMapping", "species_ucsc_matches.tsv"))
df_crossmap <- unique(df_crossmap[!is.na(ucsc_db), c("ucsc_species", "ucsc_db", "class")])

df_crossmap[, downloaded:= TRUE,]

head(df_crossmap)

NROW(df_crossmap)

species of interest:

df <- read.csv("../../validation/meta/validation_species.csv", sep = ";")
head(df)

df_full <- full_join(df_crossmap, df, by = c("ucsc_db" = "ucsc_genome", "class" = "class" ))

NROW(df_full)

setDT(df_full)

df_full[species != ucsc_species]

-> all the names make sense

head(df_full)

df_full[is.na(ucsc_species),ucsc_species:=species, by = row.names(df_full[is.na(ucsc_species)])]

df_full[ucsc_db == "canFam4", ucsc_db:="canFam5",]

wd <- file.path(analysis_dir,"validation", "02_insilico_digest")
dir.create(wd)
setwd(wd)

## Download the genome tracks for every selected genome:
Takes a while

output_dir <- file.path(data_dir, "resources", "reference_genomes")

getTrack <- function(UCSC_session, track_id, dir_out, genome_id){
    my_track <- track(ucscTableQuery(UCSC_session, track=UCSCnames[track_id]))
    export.bed(my_track, file.path(dir_out, 
                    paste0(genome_id, "_", UCSCnames[track_id],".bed")))
}

for (i in 1:NROW(df_full)){
    print(as.character(df_full[i]$ucsc_species))
    #where to save tracks
    dir_out <- file.path(output_dir, df_full[i]$ucsc_db,  "tracks")
    dir.create(dir_out, recursive = TRUE)
    
    #new rtracklayer seccion
    mySession <- browserSession()
    genome(mySession) <- df_full[i]$ucsc_db
    
    ##UCSC browser available tracks
    UCSCnames <- trackNames(ucscTableQuery(mySession))
    
    ## download CpGs:
    if(!"CpG Islands" %in% names(UCSCnames)) print(paste0("CpG islands not available for ",df_full[i]$ucsc_db))
    else getTrack(mySession, "CpG Islands", dir_out, df_full[i]$ucsc_db)
    
    ##download repeats:
    if(!"RepeatMasker" %in% names(UCSCnames)){
    print(paste0("RepeatMasker not available for ",df_full[i]$ucsc_db))
        
    if("simpleRepeat" %in% names(UCSCnames)) getTrack(mySession, "Simple Repeats", dir_out, df_full[i]$ucsc_db)
    else print("also no simpleRepeat")
    }
    else getTrack(mySession, "RepeatMasker", dir_out, df_full[i]$ucsc_db)
    
    ## gene annotation:
    possible_ids <- c("NCBI RefSeq", "RefSeq Genes","Ensembl Gene", "Other RefSeq", "AUGUSTUS", "Genscan Genes")
    k = 1
    while(!possible_ids[k] %in% names(UCSCnames) & k <= length(possible_ids)){
        k <- k+1
    }
    if(k > length(possible_ids)) print(paste0("no gene annotation available for ", df_full[i]$ucsc_db))
    else{
        getTrack(mySession, possible_ids[k], dir_out, df_full[i]$ucsc_db)
    }

   
    print(df_full[i]$ucsc_db)
}

## saving which information was avaliable for which genome:

tracks_list <- sapply(df_full$ucsc_db, function(x) 
    gsub(".bed", "", gsub(paste0(x, "_"), "",
                      list.files(file.path(output_dir, x,  "tracks")))))

#where we have nothing (need to check by hand)
which(sapply(tracks_list, function(x) length(x)==0))

##for other
colnames_tracks <- unique(unlist(tracks_list,use.names=FALSE))
tracks_present <- setNames(data.table(matrix(nrow = 0, ncol = length(colnames_tracks))),
                           colnames_tracks)

for(i in seq_along(tracks_list)){
temp_dt <- setNames(data.table(matrix(1, nrow = 1, ncol = length(tracks_list[[i]]))),
                    tracks_list[[i]])

  tracks_present <- rbind(tracks_present, temp_dt, fill = TRUE)
}

tracks_present[, ucsc_db := names(tracks_list[sapply(tracks_list, function(x) length(x) > 0)]),]


colnames(tracks_present)

##check if we have gene annotation for all
 tracks_present[is.na(refSeqComposite) & is.na(refGene) & is.na(xenoRefGene)]$ucsc_db

## repeats
 tracks_present[(is.na(rmsk) & is.na(simpleRepeat))]

## cpg islands
 tracks_present[is.na(cpgIslandExt)]

getwd()

my_wt(tracks_present, file.path(analysis_dir, "validation","02_insilico_digest","tracks_present.tsv"))

## Rerunning the concat script to download and concat genomes

write.table(df_full[is.na(downloaded), c("ucsc_db", "ucsc_species")],"genomes_ucsc_todownload_for_insilico.tsv", sep = "\t", row.names = F, col.names = F, quote = F )

## Saving the table which genomes to run the insilico digestion on

df_full[ucsc_species %in% df_full$ucsc_species[duplicated(df_full$ucsc_species)],]

df_full <- df_full[!ucsc_db %in% c("canFam3", "mm10", "petMar2")]

df_full <- df_full[! (ucsc_db == "macEug2" & class == "Mammalia")]

df_full <- df_full[! (ucsc_db == "sarHar1" & class == "Mammalia")]

my_wt(df_full[, c("ucsc_db", "ucsc_species", "species", "scientific_name", "class")], "genomes_to_run.tsv")

NROW(df_full)
